+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 <!-- pnpm 9.4.0 iwr https://get.pnpm.io/install.ps1 -useb | iex-->
 <!-- dprint 0.45.1 -->
 
+- feat: hover/Completion `__New()` `__Set()` etc...
+
 ## v0.0.62(2024-06-28)
 
 <!-- pnpm 9.4.0 iwr https://get.pnpm.io/install.ps1 -useb | iex-->